+++ model/model.go
@@ -70,8 +70,15 @@ func setToErrorState(appState *core.ApplicationState, main *core.MainView, list 
 }
 
 func resetStates(appState *core.ApplicationState, ret *handler.StoryHandler, service hn.Service) {
+	lastScreenHeight := appState.ScreenHeight
+	currentScreenHeight := screen.GetTerminalHeight()
+
 	resetApplicationState(appState)
-	ret.Reset()
+
+	if lastScreenHeight != currentScreenHeight {
+		ret.Reset()
+	}
+
 	service.Init(appState.StoriesToShow)
 }
 
@@ -96,6 +103,9 @@ func initializeView(appState *core.ApplicationState, main *core.MainView, ret *h
 }
 
 func Refresh(app *cview.Application, main *core.MainView, appState *core.ApplicationState) {
+	// Hack: Set screenHeight to 0 in order to trigger a full fetch of the submissions
+	appState.ScreenHeight = 0
+
 	afterResizeFunc := app.GetAfterResizeFunc()
 	afterResizeFunc(appState.ScreenWidth, appState.ScreenHeight)