+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.30
 _WIP_
 
+**Cosmetic**
+- Stylized `less` prompt
 
 ## 1.29
 _2021-12-14_